Rename global variable relations
This is the beginning of a series of commits that will try to unify the naming
conventions of the Fact Generator and Datalog code. It's important to undertake
this while the git history of the project is relatively short, so as to not hide
valuable history by doing it later.

At the same time, this set of changes will endeavor to make consistent,
reasonable trade-offs between identifier clarity and brevity, and document these
choices.