FactGenerator: Capture instr attr naming schemes (pt. 1)
Begin systematizing the names of attributes of instructions. This commit
endeavors not to rename anything that appears in the fact generator nor Datalog,
but merely to capture the existing system(s) of instruction attribute names.
This will make it more obvious which ones lie outside these conventions, and
make it easier to change them.

This commit ends at `trunc`, further commits will do the rest.