datalog: Fix some Soufflé warnings
Some of these can't be fixed, others were fixed by "manually unrolling"
disjunctions (which Soufflé doesn't handle properly for warnings about unused
variables).

For the assertions, the breaking up of disjunctions is a bit more verbose, but
should be both more performant (as some clauses could be removed from some
disjuncts) and more precise (since the removed clauses don't have to be true to
catch the assertion failure).

The goal is to make it easier to see and avoid introducing new warnings.