Remove subset-and-unification target
There was some idea that this could be a useful mode of analysis, but it wasn't
executed to a point where it's actually usable. The generated C++ file is huge
and takes a long time to compile, so let's avoid it.

The "debug" build still includes both analyses, ensuring compatibility (e.g.,
that they don't define relations with the same name).